Make separate account page, make new dashboard starter page, clean up / to not require auth, middleware now redirects to it
