+++ app/dashboard/account/page.tsx
@@ -1,7 +1,7 @@
 import { auth } from "@/lib/auth";
 import { headers } from "next/headers";
 import { redirect } from "next/navigation";
-import UserCard from "./user-card";
+import UserCard from "../user-card";
 import { OrganizationCard } from "./organization-card";
 import AccountSwitcher from "@/components/account-switch";
