+++ middleware.ts
@@ -15,7 +15,7 @@ export async function middleware(request: NextRequest) {
 	);
 
 	if (!session) {
-		return NextResponse.redirect(new URL("/sign-in", request.url));
+		return NextResponse.redirect(new URL("/", request.url));
 	}
 	return NextResponse.next();
 }
